refactor(ale_claw): move analyze_image + milestone tools into harness/tools/
Both are registered model-facing tools (register_tool), so they belong with
the other tool implementations, not at base. Behavior-preserving: git mv +
import-path updates (analyze_image's .utils import -> ..utils; tools.py
in-method imports -> sibling; test patch-strings -> tools.*). 928/3.
