+++ ale_run/agents/ale_claw/harness/tools/tools.py
@@ -231,8 +231,8 @@ def build_tools(
             0.10, 32 KB, 128 KB) — matches OpenClaw
             ``resolveAdaptiveReadMaxBytes``).
     """
-    from ..analyze_image import AnalyzeImageTool
-    from ..milestone import MilestoneTool
+    from .analyze_image import AnalyzeImageTool
+    from .milestone import MilestoneTool
 
     milestone_tool = MilestoneTool(session.interface)
     analyze_image_tool = AnalyzeImageTool(
